refactored. Moved to Password package
