+++ app/src/main/java/com/example/quicknotes/password/PasswordDao.java
@@ -1,7 +1,6 @@
-package com.example.quicknotes;
+package com.example.quicknotes.password;
 
 import androidx.room.Dao;
-import androidx.room.Delete;
 import androidx.room.Insert;
 import androidx.room.OnConflictStrategy;
 import androidx.room.Query;
