+++ app/src/main/java/com/example/quicknotes/password/PasswordEntity.java
@@ -1,4 +1,4 @@
-package com.example.quicknotes;
+package com.example.quicknotes.password;
 
 import androidx.room.Entity;
 import androidx.room.PrimaryKey;
