+++ app/src/main/java/com/example/quicknotes/password/PasswordFragment.java
@@ -1,4 +1,4 @@
-package com.example.quicknotes;
+package com.example.quicknotes.password;
 
 import android.annotation.SuppressLint;
 import android.os.AsyncTask;
@@ -11,18 +11,10 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
 
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.Button;
-import android.widget.EditText;
-
 import androidx.fragment.app.Fragment;
 
+import com.example.quicknotes.R;
 import com.example.quicknotes.database.AppDatabase;
-
-import java.util.Set;
 
 public class PasswordFragment extends Fragment {
